requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pillow==10.1.0
torch==2.1.0
transformers==4.35.0
accelerate==0.24.1
sentencepiece==0.1.99
protobuf==4.25.0
pydantic==2.5.0
python-multipart==0.0.6
